Implemented E2EE, storing images

+++ controllers/resourceController.js
@@ -4,12 +4,13 @@ const { User } = require("../models/schema");
 const ExpressError = require('../utils/ExpressError');
 const permitAuthorization = require('../utils/permitAuthorization');
 
-function createResource(key, value, user) {
+function createResource(key, value, user, symmetricKey) {
   return new Promise((resolve, reject) => {
     Resource.create({
       resourceName: key,
       resourceValue: value,
       resourceOwner: user._id,
+      symmetricKey: symmetricKey
     })
       .then(async (resource) => {
         const userOwnedResources = await Resource.find({
@@ -31,10 +32,10 @@ function createResource(key, value, user) {
 
 module.exports.createResource = async (req, res) => {
   try {
-    const { key, value } = req.body;
+    const { key, value, symmetricKey } = req.body;
     const user = req.userDetails;
     if (key != "" && value != "") {
-      createResource(key, value, user).then((userOwnedResources) => {
+      createResource(key, value, user, symmetricKey).then((userOwnedResources) => {
         return res.status(200).send(userOwnedResources);
       });
     } else {
@@ -54,6 +55,11 @@ module.exports.createResource = async (req, res) => {
 
 module.exports.deleteResource = async (req, res) => {
   const { resourceId } = req.body;
+  const check = await permitAuthorization.checkPermission(req.userDetails.username, "delete", resourceId, "cred");
+  if(!check.check) return res.status(403).json({
+      deleted: false,
+      msg: "User not permitted"
+  });
   Resource.deleteOne({ _id: resourceId , 
     $or:[
       {resourceOwner: req.userDetails._id },
@@ -68,21 +74,24 @@ module.exports.deleteResource = async (req, res) => {
           msg: "User not permitted"
         });
       }
-      return res.status(200).json({
-        msg: "Resource deleted successfully",
-        deleted: true
-       });
-    })
-    .catch((err) => {
-      throw new ExpressError(`Error found while deleting resource:::  ${err}`, 500);
-    });
+      User.updateMany({},{ $pull: {encryptedSymmetricKeys: { resourceId: resourceId }}})
+      .then(result => {
+        return res.status(200).json({
+          msg: "Resource deleted successfully",
+          deleted: true
+         });
+      }).catch((err) => {
+        throw new ExpressError(`Error found while deleting resource:::  ${err}`, 500);
+      });
+    }).catch((err) => {
+        throw new ExpressError(`Error found while deleting resource:::  ${err}`, 500);
+      });
 };
 
 module.exports.addMemberToResource = async (req, res) => {
-  const { addedUser, role, resourceId } = req.body;
-  if (!addedUser || !role || !resourceId)
+  const { addedUser, role, resourceId, stringEncryptedSymmetricKey } = req.body;
+  if (!addedUser || !role || !resourceId || !stringEncryptedSymmetricKey)
     return res.status(200).json({ response: 0 });
-
   const resource = await Resource.findOne({ _id: resourceId, resourceOwner: req.userDetails._id });
   if (!resource) {
     return res.json({
@@ -90,23 +99,24 @@ module.exports.addMemberToResource = async (req, res) => {
       msg: "No matching {credential} found for adding {member} ...(*￣０￣)ノ",
     });
   }
-  const user = await User.findOne({ username: addedUser });
+  const user = await User.findOneAndUpdate({ username: addedUser, 'encryptedSymmetricKeys.resourceId': { $ne: resourceId } },
+    { $push: { encryptedSymmetricKeys: {encryptedSymmetricKey: stringEncryptedSymmetricKey, resourceId: resourceId} } },
+    { new: true }
+  );
   if (!user) {
     return res.json({
       response: 3,
       msg: "No {member} found ¯¯|_(ツ)_|¯¯",
     });
   }
-  console.log(
-    `user._id (member) is: ${user._id} && resource.resourceOwner is: ${resource.resourceOwner}`
-  );
+  console.log(`user._id (member) is: ${user._id} && resource.resourceOwner is: ${resource.resourceOwner}`);
 
   // if (user._id.toString() == resource.resourceOwner.toString()) {
   //   console.log("You are the OWNER");
   //   return res.status(200).json({ response: 2 });
   // }
   const permitCheck = await permitAuthorization.checkPermission(user.username, "share", resource._id, "cred");
-  if(permitCheck) return res.status(200).json({ response: 2 });
+  if(permitCheck.check) return res.status(200).json({ response: 2 });
 
   // check for role: add, remove others
 
@@ -209,7 +219,10 @@ module.exports.editResource = async (req, res) => {
     });
   }
   const permitResult = await permitAuthorization.checkPermission(req.userDetails.username, "edit", resId, "cred");
-  if(!permitResult) return res.status(403).json({ msg: "Not permitted"});
+  if(!permitResult) return res.status(403).json({ 
+    response: 2,
+    msg: "Not permitted"
+  });
   try{
   Resource.findOneAndUpdate(
     { _id: resId , 
@@ -230,68 +243,98 @@ module.exports.editResource = async (req, res) => {
 }
 };
 
+module.exports.getSymmetricKey = async (req, res) => {
+  const resourceId = req.body.resourceId;
+  const resource = await Resource.findOne({ _id: resourceId });
+  if(!resource) return res.status(401).json({ msg: "resource not found" });
+  return res.status(200).json({
+    symmetricKeyBase64: resource.symmetricKey
+  });
+}
+
+
 
 // --------------------- shared resources operations -----------------------//
 
 module.exports.showSharedResources = async (req, res) => {
-  const resourcesSharedWithUser = await Resource.find( { $or : [
-    {editors:{$in: [req.userDetails._id]}}, 
-    {authors:{$in: [req.userDetails._id]}}, 
-    {viewers:{$in: [req.userDetails._id]}}
-  ]}).populate("resourceOwner");
-  if(!resourcesSharedWithUser) return res.status(500).json({
-    msg: "Error occured, could not fetch documents"
-  })
-  // console.log(resourcesSharedWithUser)
-  const resourcesToSend = { viewer: [], editor:[], author: [] };
-  resourcesSharedWithUser.forEach(async resource => {
-    if(resource.viewers.includes(req.userDetails._id)){
-      permitAuthorization.checkPermission(req.userDetails.username, "view", resource._id, "cred").then((response) => {
-        if(!response.check) return res.status(401).json({ msg: "unauthorized to view" });
-      })
-      resourcesToSend.viewer.push({
-        resourceName: resource.resourceName,
-        resourceValue: resource.resourceValue,
-        resourceOwner: resource.resourceOwner.username,
-        resourceOwnerName: resource.resourceOwner.name,
-        _id: resource._id
-      });
-    }else if(resource.editors.includes(req.userDetails._id)){
-      permitAuthorization.checkPermission(req.userDetails.username, "edit", resource._id, "cred").then((response) => {
-        if(!response.check) return res.status(401).json({ msg: "unauthorized to edit" });
-      })
-      resourcesToSend.editor.push({
-        resourceName: resource.resourceName,
-        resourceValue: resource.resourceValue,
-        resourceOwner: resource.resourceOwner.username,
-        resourceOwnerName: resource.resourceOwner.name,
-        _id: resource._id
-      });
-    }else{
-      permitAuthorization.checkPermission(req.userDetails.username, "delete", resource._id, "cred").then((response) => {
-        if(!response.check) return res.status(401).json({ msg: "unauthorized to author" });
-      })
-      resourcesToSend.author.push({
-        resourceName: resource.resourceName,
-        resourceValue: resource.resourceValue,
-        resourceOwner: resource.resourceOwner.username,
-        resourceOwnerName: resource.resourceOwner.name,
-        _id: resource._id
+  try {
+    const resourcesSharedWithUser = await Resource.find({
+      $or: [
+        { editors: { $in: [req.userDetails._id] } },
+        { authors: { $in: [req.userDetails._id] } },
+        { viewers: { $in: [req.userDetails._id] } },
+      ],
+    }).populate("resourceOwner");
+
+    if (!resourcesSharedWithUser) {
+      return res.status(500).json({
+        msg: "Error occurred, could not fetch documents",
       });
     }
-  });
-  return res.status(200).json({resourcesToSend});
-}
+
+    const resourcesToSend = { viewer: [], editor: [], author: [] };
+
+    // Create an array of promises
+    const promises = resourcesSharedWithUser.map(async (resource) => {
+      if (resource.viewers.includes(req.userDetails._id)) {
+        const checkViewer = await permitAuthorization.checkPermission( req.userDetails.username, "view", resource._id, "cred" );
+        const result = req.userDetails.encryptedSymmetricKeys.find(obj => obj.resourceId.toString() === resource._id.toString() );
+        resourcesToSend.viewer.push({
+          resourceName: resource.resourceName,
+          resourceValue: resource.resourceValue,
+          resourceOwner: resource.resourceOwner.username,
+          resourceOwnerName: resource.resourceOwner.name,
+          _id: resource._id,
+          encryptedSymmetricKey: result.encryptedSymmetricKey,
+        });
+        if (!checkViewer.check) resourcesToSend.viewer.length = 0;
+      } else if (resource.editors.includes(req.userDetails._id)) {
+        const checkEditor = await permitAuthorization.checkPermission( req.userDetails.username, "edit", resource._id, "cred" );
+        const result = req.userDetails.encryptedSymmetricKeys.find(obj => obj.resourceId.toString() === resource._id.toString() );
+        resourcesToSend.editor.push({
+          resourceName: resource.resourceName,
+          resourceValue: resource.resourceValue,
+          resourceOwner: resource.resourceOwner.username,
+          resourceOwnerName: resource.resourceOwner.name,
+          _id: resource._id,
+          encryptedSymmetricKey: result.encryptedSymmetricKey,
+        });
+        if (!checkEditor.check) resourcesToSend.editor.length = 0;
+      } else {
+        const checkAuthor = await permitAuthorization.checkPermission( req.userDetails.username, "delete", resource._id, "cred" );
+        const result = req.userDetails.encryptedSymmetricKeys.find(obj => obj.resourceId.toString() === resource._id.toString() );
+        resourcesToSend.author.push({
+          resourceName: resource.resourceName,
+          resourceValue: resource.resourceValue,
+          resourceOwner: resource.resourceOwner.username,
+          resourceOwnerName: resource.resourceOwner.name,
+          _id: resource._id,
+          encryptedSymmetricKey: result.encryptedSymmetricKey,
+        });
+        if (!checkAuthor.check) resourcesToSend.author.length = 0;
+      }
+    });
+
+    // Wait for all promises to resolve
+    await Promise.all(promises);
+
+    console.log("resourcesToSend before res in showSharedResources : ", resourcesToSend);
+    return res.status(200).json({ resourcesToSend });
+  } catch (error) {
+    return res.status(500).json({ msg: "Error occurred, could not fetch documents" });
+  }
+};
+
+
 
 module.exports.showResourceInfo = async (req, res) => {
   const {resId} = req.body;
-  Resource.findOne({_id:resId}).populate('viewers').populate('editors').populate('authors').then(resource =>{
+  Resource.findOne({_id:resId}).populate('viewers').populate('editors').populate('authors').then(async resource =>{
     if(!resource) return res.json({
       msg:"Resource not found (┬┬﹏┬┬)"
     })
-    permitAuthorization.checkPermission(req.userDetails.username, "share", resource._id, "cred").then((response) => {
-      if(!response.check) return res.status(401).json({ msg: "unauthorized to view the resource info" });
-    })
+    const check = await permitAuthorization.checkPermission(req.userDetails.username, "share", resource._id, "cred");
+    if(!check.check) return res.status(401).json({ msg: "unauthorized to view the resource info" });
     // console.log(resource)
     const resourceToSend = { viewers:[], editors:[], authors:[] }
     resource.viewers.forEach(res => {
@@ -315,10 +358,11 @@ module.exports.removeResourcePermission = async (req, res) => {
       msg: "User not found ━┳━ ━┳━",
       remove: false
     })
-    permitAuthorization.checkPermission(req.userDetails.username, "share", resId, "cred").then((response) => {
-      if(!response.check) return res.status(401).json({ msg: "unauthorized to remove the permissions" });
-    })
-
+    const check = await permitAuthorization.checkPermission(req.userDetails.username, "share", resId, "cred");
+    if(!check.check) return res.status(401).json({ msg: "unauthorized to remove the permissions" });
+    // await User.updateMany( {}, { $pull: { encryptedSymmetricKeys: { resourceId: resId } } } );
+    await user.updateOne({ $pull: { encryptedSymmetricKeys: { resourceId: resId } } });
+    user.save();
     // const actualRole = role.toLowerCase()+'s';
     if(role === "Viewer"){
     Resource.findOneAndUpdate({_id: resId},{ $pull : { viewers: { $in: [user._id]}}}, {new: true})
